Made updates to address various issues where it could be possible to enumerate emails of our users.
The first issue was with ‘Reset password’. Previously, an error message would have appeared if an account with the entered email address did not exist but now we have the same response as if a user with the email address did exist. This is so someone couldn’t just put in email addresses to work out who does and doesn’t have accounts.

The second issue was with the sign in and the response time. If the login credentials were invalid, the system would respond a lot quicker then if they were valid. The `go_then_wait` method extends the response time so that it is at least two seconds. This is longer then both a valid and invalid response so a nefarious user should not be able to infer if a user with the entered email address does or does not exist.